+++ src/routes/users/reset.ts
@@ -22,10 +22,11 @@ import bcrypt from "bcrypt";
 import { createToken } from "../../lib/auth/createToken";
 import { FastifyInstance, FastifyPluginOptions, FastifyRequest } from "fastify";
 import { createSession } from "../../lib/sessions/createSession";
-import { CodeSchema, EmailSchema, PasswordSchema } from "../../lib/schemas";
+import { CodeSchema, EmailSchema, PasswordSchema, RTokenSchema } from "../../lib/schemas";
 import { sha256 } from "../../lib/sha256";
 import { Verification } from "../../models/verification";
 import { RateLimitOptions } from "@fastify/rate-limit";
+import RequireReCAPTCHA from "../../plugins/requireRecaptcha";
 
 export default (
     fastify: FastifyInstance,
@@ -37,6 +38,8 @@ export default (
             email: EmailSchema,
             code: CodeSchema,
             password: PasswordSchema,
+            sameIp: Type.Optional(Type.Boolean()),
+            rtoken: RTokenSchema,
         },
         { additionalProperties: false }
     );
@@ -49,19 +52,14 @@ export default (
                 rateLimit: <RateLimitOptions>{
                     max: 5,
                     ban: 5,
-                    keyGenerator: (
-                        req: FastifyRequest<{ Body: Static<typeof schema> }>
-                    ) => {
-                        return sha256(req.body?.email);
-                    },
-                    hook: "preHandler",
                     // one day
                     timeWindow: 1000 * 60 * 60 * 24,
                 },
             },
+            preHandler: [RequireReCAPTCHA],
         },
         async (req: FastifyRequest<{ Body: Static<typeof schema> }>, res) => {
-            const { email, code, password } = req.body;
+            const { email, code, password, sameIp } = req.body;
 
             const hashedEmail = sha256(email);
 
@@ -94,7 +92,7 @@ export default (
 
             const token = createToken(fastify.jwt, user);
 
-            await createSession(user.id, token, req.headers["user-agent"], req.ip);
+            await createSession(user.id, token, req.headers["user-agent"], req.ip, sameIp);
 
             return res.send({ token });
         }