+++ src/routes/users/verify.ts
@@ -23,10 +23,11 @@ import User from "../../models/user";
 import { FastifyInstance, FastifyPluginOptions, FastifyRequest } from "fastify";
 import { agenda } from "../../lib/agenda";
 import { createSession } from "../../lib/sessions/createSession";
-import { CodeSchema, EmailSchema } from "../../lib/schemas";
+import { CodeSchema, EmailSchema, RTokenSchema } from "../../lib/schemas";
 import { sha256 } from "../../lib/sha256";
 import { Verification } from "../../models/verification";
 import { RateLimitOptions } from "@fastify/rate-limit";
+import RequireReCAPTCHA from "../../plugins/requireRecaptcha";
 
 dotenv.config();
 
@@ -40,6 +41,7 @@ export default (
             email: EmailSchema,
             code: CodeSchema,
             sameIp: Type.Optional(Type.Boolean()),
+            rtoken: RTokenSchema
         },
         { additionalProperties: false }
     );
@@ -50,15 +52,17 @@ export default (
             schema: { body: schema },
             config: {
                 rateLimit: <RateLimitOptions>{
-                    max: 10,
-                    ban: 10,
+                    max: 5,
+                    ban: 5,
                     // 1 day
                     timeWindow: 1000 * 60 * 60 * 24,
                 },
             },
+            preHandler: [RequireReCAPTCHA]
         },
         async (req: FastifyRequest<{ Body: Static<typeof schema> }>, res) => {
             const { email, code, sameIp } = req.body;
+
             const hashedEmail = sha256(email);
 
             const verificationData = (await verificationCl.findOne({